Tables and handlebars
Added three separate tables to handlebars file gamesListView.
Added code that would pass in different arrays depending on what
the game's status is.

Next step is to connect these to the files that "call" them
The ui files
And make sure to run filters, create 3 arrays based on the status
of each game, and then pass them correctly.